complete logic separation into core
